# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "QuizApp API",
        "version": "1.0.0"
    },
    "basePath": "/qapp/public/api",
    "paths": {
        "/auth": {
            "post": {
                "summary": "Authentication Users!",
                "operationId": "authenticate",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "User object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/register": {
            "post": {
                "summary": "Register Users!",
                "operationId": "register",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "User object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/dashboard": {
            "get": {
                "summary": "Get User Dashboard",
                "operationId": "dashboard",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        }
    },
    "definitions": {
        "User": {
            "type": "object",
            "xml": {
                "name": "User"
            }
        }
    }
}